refactor(boost): Refactor boost package files
Refactored boost package files to separate datastore functionality into
its own file and improve code organization. Removed unused imports and
variables to clean up the codebase.
